Move solution to the .meta package
